backend/requirements.txt
fastapi==0.136.0
openai==2.32.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.21
requests==2.33.1
uvicorn==0.45.0
